Oppdater eksisterende TabellVisning for å passe GUI utbedringer
