+++ src/frontend/Komponenter/Behandling/Inngangsvilkår/Medlemskap/InnflyttingUtflytting.tsx
@@ -5,8 +5,8 @@ import {
 } from '../../../../App/typer/personopplysninger';
 import { formaterNullableIsoDato, formaterNullableIsoÅr } from '../../../../App/utils/formatter';
 import { slåSammenTekst } from '../../../../App/utils/utils';
+import TabellVisning from '../../Tabell/TabellVisning';
 import { FlexColumnContainer } from '../../Vilkårpanel/StyledVilkårInnhold';
-import TabellVisning from '../../Vilkårpanel/TabellVisning';
 import { VilkårInfoIkon } from '../../Vilkårpanel/VilkårInformasjonKomponenter';
 
 interface Props {
